refatoração de código em features
