+++ .gitignore
@@ -53,3 +53,5 @@ dkms.conf
 
 # vscode folder
 .vscode
+
+build/
