+++ features/area/area.h
@@ -1,0 +1,6 @@
+#ifndef AREA_H
+#define AREA_H
+
+void conversao_area(int escolha);
+
+#endif
