+++ features/comprimento/comprimento.h
@@ -1,0 +1,6 @@
+#ifndef COMPRIMENTO_H
+#define COMPRIMENTO_H
+
+void conversao_comprimento(int escolha);
+
+#endif
