+++ features/dados/dados.c
@@ -1,0 +1,6 @@
+#include <stdio.h>
+#include <features/dados/dados.h>
+
+void conversao_dados(int escolha){
+    // função de conversão de Bits, bytes, kilobytes (KB), megabytes (MB), gigabytes (GB), terabytes (TB)
+}
