+++ features/dados/dados.h
@@ -1,0 +1,6 @@
+#ifndef DADOS_H
+#define DADOS_H
+
+void conversao_dados(int escolha);
+
+#endif
