+++ features/energia/energia.h
@@ -1,0 +1,6 @@
+#ifndef ENERGIA_H
+#define ENERGIA_H
+
+void conversao_energia(int escolha);
+
+#endif
