+++ features/massa/massa.h
@@ -1,0 +1,6 @@
+#ifndef MASSA_H
+#define MASSA_H
+
+void conversao_massa(int escolha);
+
+#endif
