+++ features/potencia/potencia.h
@@ -1,0 +1,6 @@
+#ifndef POTENCIA_H
+#define POTENCIA_H
+
+void conversao_potencia(int escolha);
+
+#endif
