+++ features/temperatura/temperatura.h
@@ -1,0 +1,18 @@
+#ifndef TEMPERATURA_H
+#define TEMPERATURA_H
+
+float celsius_to_fahrenheit(float c);
+
+float celsius_to_kelvin(float c);
+
+float fahrenheit_to_celsius(float f);
+
+float fahrenheit_to_kelvin(float f);
+
+float kelvin_to_celsius(float k);
+
+float kelvin_to_fahrenheit(float k);
+
+void conversao_temperatura(int origem, int destino);
+
+#endif
