+++ features/tempo/tempo.h
@@ -1,0 +1,6 @@
+#ifndef TEMPO_H
+#define TEMPO_H
+
+void conversao_tempo(int escolha);
+
+#endif
