+++ features/velocidade/velocidade.h
@@ -1,0 +1,6 @@
+#ifndef VELOCIDADE_H
+#define VELOCIDADE_H
+
+void conversao_velocidade(int escolha);
+
+#endif
