+++ features/volume/volume.h
@@ -1,0 +1,6 @@
+#ifndef VOLUME_H
+#define VOLUME_H
+
+void conversao_volume(int escolha);
+
+#endif
